fix: map type 위치 조정
