+++ app/map/page.tsx
@@ -1,24 +1,8 @@
 'use client';
 
 import { useRef, useState } from 'react';
-import { HeadBar, Map } from '@components';
+import { HeadBar, Map, MyMap } from '@components';
 import Menu from './menu';
-
-export interface LatLng {
-  La: number;
-  Ma: number;
-}
-
-export interface MyMap {
-  maps: any;
-  map: any;
-  data: {
-    drawMode: boolean;
-    dots: LatLng[];
-    dotMarkers: any[];
-    polyLine: any;
-  };
-}
 
 export default function Home() {
   const mapRef = useRef<MyMap | undefined>(undefined); // 카카오맵 객체 받아올곳
