+++ components/index.tsx
@@ -1,5 +1,6 @@
 export { default as HeadBar } from './headBar';
 export { default as Map } from './map';
+export type { MyMap, LatLng } from './map';
 export { default as Navigation } from './navigation';
 export { default as Post } from './post';
 export { default as MyCalender } from './myCalender';
